Make Version util public for add-ons to use
